refactor(schemas): migrate Phase 11 endpoint schemas — 2 files, 10 classes (#6042)
- wake_word.py: migrate WakeWordCheckRequest/Response, WakeWordConfigRequest,
  AddWakeWordRequest, ReportFeedbackRequest → schemas_system.py
  (ReportFeedbackRequest renamed → WakeWordReportFeedbackRequest to avoid
  generic-name collisions)
- skills.py: migrate SkillConfigUpdate, SkillActionRequest, UserSkillPreferences,
  SkillFeedbackRequest, SkillInstallRequest → schemas_agent.py

Also removes dead-code UserSkillPreferences import in skills.py — class is
preserved in schemas_agent.py for re-use.
