+++ autobot-backend/api/skills.py
@@ -11,13 +11,18 @@ Includes metrics and health tracking (Issue #4339).
 """
 
 import logging
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, Optional
 
 from fastapi import APIRouter, HTTPException, Query
-from pydantic import BaseModel, Field
 
 from skills.manager import SkillManager
 from skills.registry import get_skill_registry
+from api.schemas_agent import (
+    SkillActionRequest,
+    SkillConfigUpdate,
+    SkillFeedbackRequest,
+    SkillInstallRequest,
+)
 from api.schemas_code import (
     SkillConfigUpdateResponse,
     SkillExecuteResponse,
@@ -56,46 +61,6 @@ def _get_manager() -> SkillManager:
     if _manager is None:
         _manager = SkillManager()
     return _manager
-
-
-# --- Request/Response Models ---
-
-
-class SkillConfigUpdate(BaseModel):
-    """Request body for updating a skill's configuration."""
-
-    config: Dict[str, Any] = Field(..., description="Configuration values")
-
-
-class SkillActionRequest(BaseModel):
-    """Request body for executing a skill action."""
-
-    action: str = Field(..., description="Tool/action name to execute")
-    params: Dict[str, Any] = Field(
-        default_factory=dict, description="Action parameters"
-    )
-
-
-class UserSkillPreferences(BaseModel):
-    """Request body for updating user skill preferences."""
-
-    preferences: Dict[str, bool] = Field(
-        ..., description="Mapping of skill_name -> enabled"
-    )
-
-
-class SkillFeedbackRequest(BaseModel):
-    """Request body for submitting skill feedback."""
-
-    rating: int = Field(..., description="User rating (1-5)", ge=1, le=5)
-    feedback: Optional[str] = Field(None, description="Feedback text")
-
-
-class SkillInstallRequest(BaseModel):
-    """Request body for installing a skill from the catalog."""
-
-    catalog_url: str = Field(..., description="HTTP URL of the catalog endpoint")
-    repo_id: Optional[str] = Field(None, description="SkillRepo.id to link the package to")
 
 
 # --- Endpoints ---
